Display GIFs, SVGs and unknown post types as links
