+++ app/src/main/java/eu/epitech/reyditech/MainActivity.kt
@@ -3,8 +3,14 @@ package eu.epitech.reyditech
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import coil.ImageLoader
+import coil.ImageLoaderFactory
+import coil.decode.GifDecoder
+import coil.decode.ImageDecoderDecoder
+import coil.decode.SvgDecoder
+import coil.decode.VideoFrameDecoder
 
-class MainActivity : ComponentActivity() {
+class MainActivity : ComponentActivity(), ImageLoaderFactory {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -12,4 +18,16 @@ class MainActivity : ComponentActivity() {
             Reyditech()
         }
     }
+
+    override fun newImageLoader(): ImageLoader =
+        with(ImageLoader.Builder(this)) {
+            components {
+                add(ImageDecoderDecoder.Factory())
+                add(GifDecoder.Factory())
+                add(SvgDecoder.Factory())
+                add(VideoFrameDecoder.Factory())
+            }
+            crossfade(true)
+            build()
+        }
 }
