+++ app/src/main/java/eu/epitech/reyditech/screens/MainScreen.kt
@@ -40,7 +40,7 @@ internal fun MainScreen(
             ListingPagingSource { before, after, count, limit ->
                 loginViewModel.request {
                     posts(
-                        subreddit = "feedthebeast",
+                        subreddit = "gifs",
                         type = PostType.BEST,
                         before = before,
                         after = after,
